ref(supergroups): Consolidate supergroups files into dedicated package
Move scattered supergroups modules into src/sentry/seer/supergroups/ package
and mirror the structure in tests/. Add CODEOWNERS entries for
issue-detection-backend and issue-detection-frontend teams.
